modules/backend/pyproject.toml
[project]
name = "papita-transactions-api"
version = "1.0.0"
description = "Backend supporting the RESTful API to interact with the Papita Transactions data model"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Elmorralito", email = "drestrepohinc@gmail.com" }
]
dependencies = [
    "python-dateutil (>=2.9.0,<3.0.0)",
    "papita-transactions-model (>=1.0.0,<2.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pydantic-core (>=2.28.2,<3.0.0)",
    "pydantic-extra-types (>=2.1.1,<3.0.0)",
    "argon2-cffi (>=24.1.0,<26.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "httpx (>=0.28.1,<0.30.0)",
]

[project.urls]
HomePage = "https://github.com/Elmorralito/save-ma-money/tree/main/modules/backend"

[tool.poetry]
packages = [
    { include = "papita_txnsapi", from = "src/" },
]

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
]
build-backend = "poetry.core.masonry.api"
